Moves AppStartup widget to its own file
Extracts the AppStartup widget into its own file for better code organization and maintainability. This change promotes modularity and improves the overall structure of the application.
